backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
alembic==1.14.0
asyncpg==0.30.0
redis==5.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
pydantic==2.10.0
pydantic-settings==2.6.1
httpx==0.27.2
celery==5.4.0
python-dotenv==1.0.1
emails==0.6
openpyxl==3.1.5
pandas==2.2.3
paramiko==3.5.0
